+++ utils/getUser.ts
@@ -23,6 +23,7 @@ export async function getUser() {
       id: user.id,
       name: user.name,
       email: user.email,
+      role: user.role,
       isVerified: user.isVerified,
     };
   } catch (error) {